Polytechnique Montréal/Project 2 - Web Application/client/src/app/classes/tool/tool.spec.ts
import { Vec2 } from '@app/classes/interfaces/vec2';
import { Tool } from '@app/classes/tool/tool';
import { canvasTestHelper } from '@app/classes/utils/canvas-test-helper';
import { MouseButton } from '@app/declarations/mouse.enum';
import { DrawingService } from '@app/services/drawing/drawing.service';
// tslint:disable:no-string-literal
// tslint:disable:no-magic-numbers
class ToolTest extends Tool {
    constructor(drawingService: DrawingService) {
        super(drawingService);
    }
}

describe('Tool', () => {
    let tool: ToolTest;
    let drawServiceSpy: jasmine.SpyObj<DrawingService>;
    let mouseEvent: MouseEvent;
    let pointerEvent: PointerEvent;
    let canvasStub: HTMLCanvasElement;
    let canvasContext: CanvasRenderingContext2D;

    beforeEach(() => {
        canvasStub = canvasTestHelper.canvas;
        canvasContext = canvasStub.getContext('2d') as CanvasRenderingContext2D;

        drawServiceSpy = {
            ...jasmine.createSpyObj('DrawingService', ['']),
            canvasWidth: 100,
            canvasHeight: 100,
            baseCtx: canvasContext,
        } as jasmine.SpyObj<DrawingService>;

        tool = new ToolTest(drawServiceSpy);

        mouseEvent = {
            offsetX: 25,
            offsetY: 25,
            button: MouseButton.Left,
        } as MouseEvent;

        pointerEvent = {} as PointerEvent;
    });

    it('should be created', () => {
        expect(tool).toBeTruthy();
    });

    it(' #getPositionFromMouse should get the right mouseDownCoord', () => {
        // Arrange
        const expectedResult: Vec2 = { x: 25, y: 25 };

        // Act
        const position = tool['getPositionFromMouse'](mouseEvent);

        // Assert
        expect(position).toEqual(expectedResult);
    });

    it(' #isReallyOut should return true if mouse position is out of the canvas', () => {
        // Arrange
        mouseEvent = {
            offsetX: 225,
            offsetY: 225,
            button: MouseButton.Left,
        } as MouseEvent;

        // Act
        const result = tool['isReallyOut'](mouseEvent);

        // Assert
        expect(result).toEqual(true);
    });

    it(' #isReallyOut should return false if mouse position is in the canvas', () => {
        // Arrange

        // Act
        const result = tool['isReallyOut'](mouseEvent);

        // Assert
        expect(result).toEqual(false);
    });

    /*
    Equipe 201:
    Bonjour,
    On a fait ce test pour avoir 100% de covergae.
    On sait que ce n'est pas une bonne pratique de faire des dummy tests.
    Mais c'est la seul facon de "cover" la abstract empty class
    Cordialement
  */
    it(' dummy test to cover code for abstract class (Tool)', () => {
        // Arrange
        const keyevent = {
            key: 'Shift',
        } as KeyboardEvent;

        const wheelEvent = {
            offsetX: 25,
            offsetY: 25,
            button: MouseButton.Left,
        } as WheelEvent;

        // Act
        tool.onWheelScroll(wheelEvent);
        tool.onMouseClick(mouseEvent);
        tool.onMouseDoubleClick(mouseEvent);
        tool.onMouseDown(pointerEvent);
        tool.onMouseEnter(mouseEvent);
        tool.onMouseMove(pointerEvent);
        tool.onMouseOut(mouseEvent);
        tool.onMouseUp(pointerEvent);
        tool.onKeyDown(keyevent);
        tool.onKeyUp(keyevent);
        tool.onKeyPress(keyevent);
        tool.startDrawing();
        tool.stopDrawing();

        // Assert
        expect(true).toBe(true);
    });
});
